requirements.txt
numpy==1.23.4
pandas==1.5.1
mysql-connector-python==8.0.31
python-decouple==3.6